Readme FastEdi73
- 👋 Hi, I’m @FastEdi73
- 👀 I’m interested in in Model Trains and also computers, I use to work for a major bank as communications Technician and
later worked IT. It took me a while to find out what IT was. I started at 19 in the US Army as an Electronics technician
working on Microwave and Encryption machines. When I left the military I worked for Associated Press as a technician on wirefoto
and teletype machines, and later on computers. I moved pver to Time Inc, and worked on the Photo Art work and their computers for 
working on the magazine.. I guess you can say I worked on Teletype all the way to Cisco switches (500 port) and routers.

- 🌱 I’m currently learning how to program arduino and Pic chips, I played with different proamming languages, my prefered 
one is C. But I have tried out others. 
- 💞️ I’m looking to collaborate on I have no Idea on this right now, but when I figure it out I'll be asking and looking 
for help or someone to work with.

- 📫 How to reach me ecaudet@live.com or eddieaudet47@gmail.com

<!---